move file to follow convention structure
